// src/app/models/Reporteador.ts
export class Reporteador{
   

    //RECIBO MASIVO

    idEmpresa: number | null = null;
	idSede: number | null = null;
	idCiclo: number | null = null;
	idSucursal: number | null = null;
	anio: string | null = null;
	mes: number | null = null;
	usuarioCreacion: string | null = null;


     localidad: string | null = null;
	 sector: string | null = null;
	 idTarifa: number | null = null;
	 nombreTarifa: string | null = null;
	 idSector: number | null = null;
	 usuarios: number | null = null;
	 volumenFacturado: number | null = null;
	 impAgua: number | null = null;
	 impAlcantarillado: number | null = null;
	 impCargoFijo: number | null = null;
	 impInteres: number | null = null;
	 impOtrosConceptos: number | null = null;
	 impIGV: number | null = null;
	 redondeoAnterior: number | null = null;
	 redondeoActual: number | null = null;
	 impTotalMes: number | null = null;


    
}